+++ lib/src/bones_api_entity_db_sql.dart
@@ -644,6 +644,29 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
               )
               .toList();
 
+      var missingReferenceIndexesSQLs =
+          repositoriesChecksErrors
+              .where((e) => e.missingReferenceIndexes?.isNotEmpty ?? false)
+              .expand((e) => e.generateMissingReferenceIndexesSQLs(this))
+              .expand(
+                (e) => e.buildAllSQLs(ifNotExists: true, multiline: false),
+              )
+              .toList();
+
+      var missingRelationshipReferenceIndexesSQLs =
+          repositoriesChecksErrors
+              .where(
+                (e) =>
+                    e.missingRelationshipReferenceIndexes?.isNotEmpty ?? false,
+              )
+              .expand(
+                (e) => e.generateMissingRelationshipReferenceIndexesSQLs(this),
+              )
+              .expand(
+                (e) => e.buildAllSQLs(ifNotExists: true, multiline: false),
+              )
+              .toList();
+
       var missingUniqueConstraintsSQLs =
           repositoriesChecksErrors
               .where((e) => e.missingUniqueConstraints?.isNotEmpty ?? false)
@@ -677,6 +700,10 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
         ...missingReferenceColumnsSQLs,
         '',
         ...missingReferenceConstraintsSQLs,
+        '',
+        ...missingReferenceIndexesSQLs,
+        '',
+        ...missingRelationshipReferenceIndexesSQLs,
         '',
         ...missingUniqueConstraintsSQLs,
         '',
@@ -849,6 +876,9 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
             .toList();
 
     var missingFieldReferenceConstraints = <String>{};
+    var missingFieldReferenceIndexes = <String>{};
+    var missingRelationshipFieldReferenceIndexes =
+        <(String, String, TypeInfo)>{};
 
     var missingUniqueConstraints = <String>{};
 
@@ -903,6 +933,67 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
             fieldEntityTypes.keys
                 .where((f) => !fieldsReferences.containsKey(f))
                 .toSet();
+
+        if (!dialect.foreignKeyCreatesImplicitIndex) {
+          missingFieldReferenceIndexes =
+              fieldsReferencedTables.entries
+                  .where((e) => e.value.indexName == null)
+                  .map((e) => e.key)
+                  .toSet();
+        }
+      }
+
+      // RELATIONSHIP CONSTRAINTS //
+
+      var fieldListEntityTypes =
+          entityHandler
+              .getFieldsListEntityTypes()
+              .entries
+              .map((e) {
+                var entityType = e.value.entityType;
+                if (entityType == null) return null;
+
+                var refRepository = repository.provider
+                    .getEntityRepositoryByType(entityType);
+
+                if (refRepository == null ||
+                    !refRepository.isSameEntityManager(repository)) {
+                  return null;
+                }
+
+                return MapEntry(e.key, entityType);
+              })
+              .nonNulls
+              .toMapFromEntries();
+
+      if (fieldListEntityTypes.isNotEmpty) {
+        if (!dialect.foreignKeyCreatesImplicitIndex) {
+          missingRelationshipFieldReferenceIndexes =
+              collectionReferenceFields.entries
+                  .expand((e) {
+                    var relFieldRef = e.value;
+                    if (relFieldRef == null) {
+                      return <(String, String, TypeInfo)>[];
+                    }
+
+                    return [
+                      if (relFieldRef.sourceRelationshipFieldIndex == null)
+                        (
+                          relFieldRef.relationshipTable,
+                          relFieldRef.sourceRelationshipField,
+                          relFieldRef.sourceFieldEntityType,
+                        ),
+                      if (relFieldRef.targetRelationshipFieldIndex == null)
+                        (
+                          relFieldRef.relationshipTable,
+                          relFieldRef.targetRelationshipField,
+                          relFieldRef.targetFieldEntityType,
+                        ),
+                    ];
+                  })
+                  .nonNulls
+                  .toSet();
+        }
       }
 
       // UNIQUE CONSTRAINTS //
@@ -982,6 +1073,8 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
         missingReferenceColumns.isNotEmpty ||
         missingCollectionReferenceColumns.isNotEmpty ||
         missingFieldReferenceConstraints.isNotEmpty ||
+        missingFieldReferenceIndexes.isNotEmpty ||
+        missingRelationshipFieldReferenceIndexes.isNotEmpty ||
         missingUniqueConstraints.isNotEmpty ||
         missingEnumConstraints.isNotEmpty ||
         missingEnumValues.isNotEmpty) {
@@ -991,6 +1084,8 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
         "${missingReferenceColumns.isNotEmpty ? '-- missingReferenceColumns: $missingReferenceColumns\n' : ''}"
         "${missingCollectionReferenceColumns.isNotEmpty ? '-- missingCollectionReferenceColumns: $missingCollectionReferenceColumns\n' : ''}"
         "${missingFieldReferenceConstraints.isNotEmpty ? '-- missingFieldReferenceConstraints: $missingFieldReferenceConstraints\n' : ''}"
+        "${missingFieldReferenceIndexes.isNotEmpty ? '-- missingFieldReferenceIndexes: $missingFieldReferenceIndexes\n' : ''}"
+        "${missingRelationshipFieldReferenceIndexes.isNotEmpty ? '-- missingRelationshipFieldReferenceIndexes: $missingRelationshipFieldReferenceIndexes\n' : ''}"
         "${missingUniqueConstraints.isNotEmpty ? '-- missingUniqueConstraints: $missingUniqueConstraints\n' : ''}"
         "${missingEnumConstraints.isNotEmpty ? '-- missingEnumConstraints: $missingEnumConstraints\n' : ''}"
         "${missingEnumValues.isNotEmpty ? '-- missingEnumValues: $missingEnumValues\n' : ''}",
@@ -1010,6 +1105,12 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
             .toList(),
         missingFieldReferenceConstraints
             .map((f) => _DBTableColumn.fromEntityHandler(entityHandler, f))
+            .toList(),
+        missingFieldReferenceIndexes
+            .map((f) => _DBTableColumn.fromEntityHandler(entityHandler, f))
+            .toList(),
+        missingRelationshipFieldReferenceIndexes
+            .map((f) => _DBRelationshipTableColumn(f.$1, f.$2, f.$3, []))
             .toList(),
         missingUniqueConstraints
             .map((f) => _DBTableColumn.fromEntityHandler(entityHandler, f))
@@ -1033,7 +1134,7 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
     return _DBTableCheck.ok();
   }
 
-  MapEntry<String, TableRelationshipReference?>?
+  MapEntry<String, TableRelationshipReferenceEntityTyped?>?
   _checkDBTableSchemeReferenceField(
     EntityHandler<Object> entityHandler,
     TableScheme scheme,
@@ -1058,13 +1159,18 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
     var targetTable = getTableForType(entityType);
     if (targetTable == null) return null;
 
-    TableRelationshipReference? tableRef;
+    TableRelationshipReferenceEntityTyped? tableRefEntityTyped;
 
     try {
-      tableRef = scheme.getTableRelationshipReference(
+      var tableRef = scheme.getTableRelationshipReference(
         sourceTable: table,
         sourceField: fieldName,
         targetTable: targetTable,
+      );
+
+      tableRefEntityTyped = tableRef?.copyWithEntityTypes(
+        entityHandler.typeInfo,
+        entityType,
       );
     } catch (e) {
       _log.warning(
@@ -1073,7 +1179,7 @@ abstract class DBSQLAdapter<C extends Object> extends DBRelationalAdapter<C>
       );
     }
 
-    return MapEntry(fieldName, tableRef);
+    return MapEntry(fieldName, tableRefEntityTyped);
   }
 
   bool checkDBTableField(
@@ -3311,6 +3417,10 @@ class _DBTableCheck {
 
   List<_DBTableColumn>? missingReferenceConstraints;
 
+  List<_DBTableColumn>? missingReferenceIndexes;
+
+  List<_DBRelationshipTableColumn>? missingRelationshipReferenceIndexes;
+
   List<_DBTableColumn>? missingUniqueConstraints;
 
   List<_DBTableColumn>? missingEnumConstraints;
@@ -3323,6 +3433,8 @@ class _DBTableCheck {
     this.missingReferenceColumns,
     this.missingCollectionReferenceColumns,
     this.missingReferenceConstraints,
+    this.missingReferenceIndexes,
+    this.missingRelationshipReferenceIndexes,
     this.missingUniqueConstraints,
     this.missingEnumConstraints,
     this.missingEnumValues,
@@ -3340,6 +3452,8 @@ class _DBTableCheck {
       (missingReferenceColumns?.isEmpty ?? true) &&
       (missingCollectionReferenceColumns?.isEmpty ?? true) &&
       (missingReferenceConstraints?.isEmpty ?? true) &&
+      (missingReferenceIndexes?.isEmpty ?? true) &&
+      (missingRelationshipReferenceIndexes?.isEmpty ?? true) &&
       (missingUniqueConstraints?.isEmpty ?? true) &&
       (missingEnumConstraints?.isEmpty ?? true) &&
       (missingEnumValues?.isEmpty ?? true);
@@ -3347,16 +3461,16 @@ class _DBTableCheck {
   bool get isError => !isOK;
 
   List<AlterTableSQL> generateMissingColumnsSQLs(SQLGenerator sqlGenerator) =>
-      _generateMissingSQLs(missingColumns, sqlGenerator);
+      _generateMissingColumnSQLs(missingColumns, sqlGenerator);
 
   List<AlterTableSQL> generateMissingReferenceColumnsSQLs(
     SQLGenerator sqlGenerator,
-  ) => _generateMissingSQLs(missingReferenceColumns, sqlGenerator);
+  ) => _generateMissingColumnSQLs(missingReferenceColumns, sqlGenerator);
 
   List<AlterTableSQL> generateMissingReferenceConstraintsSQLs(
     SQLGenerator sqlGenerator,
   ) {
-    var columnsSQLs = _generateMissingSQLs(
+    var columnsSQLs = _generateMissingColumnSQLs(
       missingReferenceConstraints,
       sqlGenerator,
     );
@@ -3365,6 +3479,34 @@ class _DBTableCheck {
         columnsSQLs.expand((e) => e.constraints ?? <AlterTableSQL>[]).toList();
 
     return constraints;
+  }
+
+  List<CreateIndexSQL> generateMissingReferenceIndexesSQLs(
+    SQLGenerator sqlGenerator,
+  ) {
+    var columnsSQLs = _generateMissingColumnSQLs(
+      missingReferenceIndexes,
+      sqlGenerator,
+    );
+
+    var indexes =
+        columnsSQLs.expand((e) => e.indexes ?? <CreateIndexSQL>[]).toList();
+
+    return indexes;
+  }
+
+  List<CreateIndexSQL> generateMissingRelationshipReferenceIndexesSQLs(
+    SQLGenerator sqlGenerator,
+  ) {
+    var columnsSQLs = _generateMissingColumnSQLs(
+      missingRelationshipReferenceIndexes,
+      sqlGenerator,
+    );
+
+    var indexes =
+        columnsSQLs.expand((e) => e.indexes ?? <CreateIndexSQL>[]).toList();
+
+    return indexes;
   }
 
   List<AlterTableSQL> generateMissingUniqueConstraintsSQLs(
@@ -3386,7 +3528,7 @@ class _DBTableCheck {
     sqlGenerator,
   );
 
-  List<AlterTableSQL> _generateMissingSQLs(
+  List<AlterTableSQL> _generateMissingColumnSQLs(
     List<_DBTableColumn>? missing,
     SQLGenerator sqlGenerator,
   ) {
@@ -3396,16 +3538,18 @@ class _DBTableCheck {
     }
 
     var sqls =
-        missing
-            .map(
-              (f) => sqlGenerator.generateAddColumnAlterTableSQL(
-                scheme.name,
-                f.name,
-                f.type,
-                entityFieldAnnotations: f.annotations,
-              ),
-            )
-            .toList();
+        missing.map((f) {
+          var schemeName =
+              f is _DBRelationshipTableColumn
+                  ? f.relationshipTable
+                  : scheme.name;
+          return sqlGenerator.generateAddColumnAlterTableSQL(
+            schemeName,
+            f.name,
+            f.type,
+            entityFieldAnnotations: f.annotations,
+          );
+        }).toList();
 
     return sqls;
   }
@@ -3461,9 +3605,17 @@ class _DBTableCheck {
     var missingCollectionReferenceColumns =
         this.missingCollectionReferenceColumns ?? [];
 
+    var missingReferenceConstraints = this.missingReferenceConstraints ?? [];
+    var missingReferenceIndexes = this.missingReferenceIndexes ?? [];
+    var missingRelationshipReferenceIndexes =
+        this.missingRelationshipReferenceIndexes ?? [];
+
     if (missingColumns.isNotEmpty ||
         missingReferenceColumns.isNotEmpty ||
-        missingCollectionReferenceColumns.isNotEmpty) {
+        missingCollectionReferenceColumns.isNotEmpty ||
+        missingReferenceConstraints.isNotEmpty ||
+        missingReferenceIndexes.isNotEmpty ||
+        missingRelationshipReferenceIndexes.isNotEmpty) {
       var repoType = repository?.type;
       var schemeTableName = scheme?.name;
 
@@ -3474,6 +3626,19 @@ class _DBTableCheck {
       var missingReferenceColumnsStr = missingReferenceColumns
           .map((e) => '${e.type.toString(withT: false)} ${e.name}')
           .join(' , ');
+
+      var missingReferenceConstraintsStr = missingReferenceConstraints
+          .map((e) => '${e.type.toString(withT: false)} ${e.name}')
+          .join(' , ');
+
+      var missingReferenceIndexesStr = missingReferenceIndexes
+          .map((e) => '${e.type.toString(withT: false)} ${e.name}')
+          .join(' , ');
+
+      var missingRelationshipReferenceIndexesStr =
+          missingRelationshipReferenceIndexes
+              .map((e) => '${e.type.toString(withT: false)} ${e.name}')
+              .join(' , ');
 
       var missingCollectionReferenceColumnsStr =
           missingCollectionReferenceColumns
@@ -3487,6 +3652,12 @@ class _DBTableCheck {
           "  -- missingColumns: [$missingColumnsStr]",
         if (missingReferenceColumns.isNotEmpty)
           "  -- missingReferenceColumns: [$missingReferenceColumnsStr]",
+        if (missingReferenceConstraints.isNotEmpty)
+          "  -- missingReferenceConstraints: [$missingReferenceConstraintsStr]",
+        if (missingReferenceIndexes.isNotEmpty)
+          "  -- missingReferenceIndexes: [$missingReferenceIndexesStr]",
+        if (missingRelationshipReferenceIndexes.isNotEmpty)
+          "  -- missingRelationshipReferenceIndexes: [$missingRelationshipReferenceIndexesStr]",
         if (missingCollectionReferenceColumns.isNotEmpty)
           "  -- missingCollectionReferenceColumns: [$missingCollectionReferenceColumnsStr]",
       ].join('\n');
@@ -3496,6 +3667,20 @@ class _DBTableCheck {
 
     return error?.toString() ?? '';
   }
+}
+
+class _DBRelationshipTableColumn extends _DBTableColumn {
+  final String relationshipTable;
+
+  _DBRelationshipTableColumn(
+    this.relationshipTable,
+    super.name,
+    super.type,
+    super.annotations,
+  );
+
+  @override
+  String toString() => '{$relationshipTable.$name: $type}';
 }
 
 class _DBTableColumn {