+++ lib/src/bones_api_sql_builder.dart
@@ -105,6 +105,9 @@ class SQLDialect extends DBDialect {
   /// If `true` indicates that the `VARCHAR` can be defined without a maximum size.
   final bool acceptsVarcharWithoutMaximumSize;
 
+  /// If `true`, foreign keys implicitly create an index on the referencing columns.
+  final bool foreignKeyCreatesImplicitIndex;
+
   const SQLDialect(
     super.name, {
     this.elementQuote = '',
@@ -115,6 +118,7 @@ class SQLDialect extends DBDialect {
     this.acceptsInsertIgnore = false,
     this.acceptsInsertOnConflict = false,
     this.acceptsVarcharWithoutMaximumSize = false,
+    this.foreignKeyCreatesImplicitIndex = true,
   });
 
   @override
@@ -129,7 +133,17 @@ class SQLDialect extends DBDialect {
 
   @override
   String toString() {
-    return 'SQLDialect{name: $name, elementQuote: $elementQuote, acceptsOutputSyntax: $acceptsOutputSyntax, acceptsReturningSyntax: $acceptsReturningSyntax, acceptsTemporaryTableForReturning: $acceptsTemporaryTableForReturning, acceptsInsertDefaultValues: $acceptsInsertDefaultValues, acceptsInsertIgnore: $acceptsInsertIgnore, acceptsInsertOnConflict: $acceptsInsertOnConflict}';
+    return 'SQLDialect{'
+        'name: $name,'
+        'elementQuote: $elementQuote, '
+        'acceptsOutputSyntax: $acceptsOutputSyntax, '
+        'acceptsReturningSyntax: $acceptsReturningSyntax, '
+        'acceptsTemporaryTableForReturning: $acceptsTemporaryTableForReturning, '
+        'acceptsInsertDefaultValues: $acceptsInsertDefaultValues, '
+        'acceptsInsertIgnore: $acceptsInsertIgnore, '
+        'acceptsInsertOnConflict: $acceptsInsertOnConflict, '
+        'foreignKeyCreatesImplicitIndex: $foreignKeyCreatesImplicitIndex'
+        '}';
   }
 }
 
@@ -1610,6 +1624,15 @@ abstract mixin class SQLGenerator {
       );
 
       constraints.add(AlterTableSQL(dialect, table, [constraintEntry]));
+
+      if (!dialect.foreignKeyCreatesImplicitIndex) {
+        var indexName = '${table}__${columnName}__idx';
+
+        indexes ??= [];
+        indexes.add(
+          CreateIndexSQL(dialect, table, columnName, indexName, q: q),
+        );
+      }
     }
 
     var alterTableSQL = AlterTableSQL(
@@ -1920,6 +1943,13 @@ abstract mixin class SQLGenerator {
           ],
         ),
       );
+
+      if (!dialect.foreignKeyCreatesImplicitIndex) {
+        var indexName = '${table}__${columnName}__idx';
+        indexSQLs.add(
+          CreateIndexSQL(dialect, table, columnName, indexName, q: q),
+        );
+      }
     }
 
     sqlEntries.sort((a, b) {
@@ -2049,12 +2079,25 @@ abstract mixin class SQLGenerator {
         ),
       ];
 
+      var relIndexes = <CreateIndexSQL>[];
+
+      if (!dialect.foreignKeyCreatesImplicitIndex) {
+        var indexName1 = '${relName}__${srcFieldName}__idx';
+        var indexName2 = '${relName}__${dstFieldName}__idx';
+
+        relIndexes.addAll([
+          CreateIndexSQL(dialect, relName, srcFieldName, indexName1, q: q),
+          CreateIndexSQL(dialect, relName, dstFieldName, indexName2, q: q),
+        ]);
+      }
+
       var relSQL = CreateTableSQL(
         dialect,
         relName,
         sqlRelEntries,
         q: q,
         parentTable: table,
+        indexes: relIndexes,
       );
       relationshipSQLs.add(relSQL);
     }